Change bandwidth type from int to long.
It's required to change bandwidth type from int to long
to have a consistent type in different DTO's.

Closes #928

+++ services/src/atdd/src/test/java/org/openkilda/atdd/FlowCrudBasicRunTest.java
@@ -106,7 +106,7 @@ public class FlowCrudBasicRunTest {
     @Then("^flow (.*) with (.*) (\\d+) (\\d+) and (.*) (\\d+) (\\d+) and (\\d+) could be created$")
     public void checkFlowCreation(final String flowId, final String sourceSwitch, final int sourcePort,
                                   final int sourceVlan, final String destinationSwitch, final int destinationPort,
-                                  final int destinationVlan, final int bandwidth) throws Exception {
+                                  final int destinationVlan, final long bandwidth) throws Exception {
         Flow expectedFlow = new Flow(FlowUtils.getFlowName(flowId), bandwidth, false, 0, flowId, null, sourceSwitch,
                 destinationSwitch, sourcePort, destinationPort, sourceVlan, destinationVlan, 0, 0, null, null);
