+++ services/src/pce/src/main/java/org/openkilda/pce/api/FlowAdapter.java
@@ -1,13 +1,26 @@
+/* Copyright 2018 Telstra Open Source
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
 package org.openkilda.pce.api;
 
-import org.neo4j.driver.v1.Record;
 import org.openkilda.messaging.Utils;
 import org.openkilda.messaging.info.event.PathInfoData;
 import org.openkilda.messaging.model.Flow;
 import org.openkilda.messaging.payload.flow.FlowState;
-import org.openkilda.pce.provider.NeoDriver;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+
+import org.neo4j.driver.v1.Record;
 
 import java.io.IOException;
 
@@ -17,7 +30,7 @@ public class FlowAdapter {
     public FlowAdapter(Record dbRecord) {
         String pathJson = dbRecord.get("path").asString().trim();
 
-        if (pathJson.equals("null")){
+        if (pathJson.equals("null")) {
             pathJson = "{\"path\": [], \"latency_ns\": 0, \"timestamp\": 0}";
         }
 
@@ -25,34 +38,35 @@ public class FlowAdapter {
          * The 'clazz' value is stripped when storing in the database, but we need it in the string
          * in order for MAPPER to do its thing.  So, let's add it back in at the very beginning.
          */
-        String start = pathJson.substring(0,pathJson.length()-1);
+        String start = pathJson.substring(0, pathJson.length() - 1);
         PathInfoData path;
-        pathJson = start+", \"clazz\":\"org.openkilda.messaging.info.event.PathInfoData\"}";
+        pathJson = start + ", \"clazz\":\"org.openkilda.messaging.info.event.PathInfoData\"}";
 
         try {
             path = Utils.MAPPER.readValue(pathJson, PathInfoData.class);
         } catch (IOException e) {
-            throw new IllegalArgumentException(String.format(
-                    "Can\'t deserialize flow path: json=%s", pathJson), e);
+            throw new IllegalArgumentException(
+                    String.format("Can\'t deserialize flow path: json=%s", pathJson), e);
         }
 
-        flow = new Flow(
-                dbRecord.get(Utils.FLOW_ID).asString(),
-                dbRecord.get("bandwidth").asInt(),
-                dbRecord.get("ignore_bandwidth").asBoolean(),
-                dbRecord.get("cookie").asLong(),
-                dbRecord.get("description").asString(),
-                dbRecord.get("last_updated").asString(),
-                dbRecord.get("src_switch").asString(),
-                dbRecord.get("dst_switch").asString(),
-                dbRecord.get("src_port").asInt(),
-                dbRecord.get("dst_port").asInt(),
-                dbRecord.get("src_vlan").asInt(),
-                dbRecord.get("dst_vlan").asInt(),
-                dbRecord.get("meter_id").asInt(),
-                dbRecord.get("transit_vlan").asInt(),
-                path, FlowState.CACHED
-        );
+        flow =
+                new Flow(
+                        dbRecord.get(Utils.FLOW_ID).asString(),
+                        dbRecord.get("bandwidth").asLong(),
+                        dbRecord.get("ignore_bandwidth").asBoolean(),
+                        dbRecord.get("cookie").asLong(),
+                        dbRecord.get("description").asString(),
+                        dbRecord.get("last_updated").asString(),
+                        dbRecord.get("src_switch").asString(),
+                        dbRecord.get("dst_switch").asString(),
+                        dbRecord.get("src_port").asInt(),
+                        dbRecord.get("dst_port").asInt(),
+                        dbRecord.get("src_vlan").asInt(),
+                        dbRecord.get("dst_vlan").asInt(),
+                        dbRecord.get("meter_id").asInt(),
+                        dbRecord.get("transit_vlan").asInt(),
+                        path,
+                        FlowState.CACHED);
     }
 
     public Flow getFlow() {
